gestion de erreurs sur endpoint du module cours

+++ src/models/cours/routes/deleteCours.js
@@ -6,21 +6,23 @@ module.exports = (app) => {
     app.delete('/api/udemy/cours/:id', (req, res) => {
 
         const id = req.params.id
-        Cour.findByPk(id)
-            .then(cour => {
-                const deletedCour = cour
-                Cour.destroy({
-                    where: { id: id }
-                })
-                .then(_ => {
-                    res.json({ message: `cour \' ${deletedCour.title} \' deleted successfully`, data: deletedCour })
+        return Cour.findByPk(id)
+                .then(cour => {
+                    if(cour === null){
+                        const message = `Il n'existe pas de cour avec l'ID ${req.params.id}`
+                        return res.status(404).json({ message })
+                    }
+                    const deletedCour = cour
+                    Cour.destroy({
+                        where: { id: id }
+                    })
+                    .then(_ => {
+                        res.json({ message: `cour \' ${deletedCour.title} \' deleted successfully`, data: deletedCour })
+                    })
                 })
                 .catch(err => {
-                    res.json({ message: 'error founded', error: err.message })
+                    const message = 'Nous avons rencontre une erreur lors de la suppresionn de cette formation, veuillez reessayer plus tard...'
+                    res.status(500).json({ message })
                 })
-            })
-            .catch(err => {
-                res.json({ message: 'error founded', error: err.message })
-            })
-    })
+        })
 }
